Move butler context manager up to LimitedButler
This allows all butlers to use it but with close() defaulting
to a no-op.
